refactor(content_ui): consolidate tag ownership
