+++ frontend/plugins/content_ui/src/modules/cms/tags/graphql/queries.ts
@@ -48,6 +48,7 @@ export const CMS_TAGS = gql`
         startCursor
         endCursor
       }
+      totalCount
     }
   }
 `;
